+++ Tests/SourceKittenFrameworkTests/SourceKitTests.swift
@@ -127,6 +127,17 @@ class SourceKitTests: XCTestCase {
         let actual = sourcekitStrings(startingWith: "source.decl.attribute.")
             .subtracting(attributesFoundInSwift5ButWeIgnore)
 
+#if compiler(>=5.2)
+        // removed in Swift 5.2
+        expected.subtract([.implicitlyUnwrappedOptional])
+#else
+        // added in Swift 5.2
+        expected.subtract([
+            .differentiable, ._nonEphemeral, ._originallyDefinedIn, ._inheritsConvenienceInitializers,
+            ._hasMissingDesignatedInitializers
+        ])
+#endif
+
 #if compiler(>=5.0)
         // removed in Swift 5.0
         expected.subtract([.silStored, .effects])