+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * Update Yams to 3.0.0.  
   [Keith Smiley](https://github.com/keith)
 
+* Add `SwiftDeclarationAttributeKind` values introduced in Swift 5.2.  
+  [JP Simard](https://github.com/jpsim)
+
 ##### Bug Fixes
 
 * Stop the program immediately if the `xcodebuild` command fails.